Verilog parsing (#21)
* verilog parser update in progress

* updating to latest verilog parser

* missed include in installation

* missing .h in installation

* code cleaning and include installation

* installation again

* missing include again

* Link SNL verilog lib to SNL primitives lib

* update FindNaja macro

* verilog parsing in progress

* Dumping constants in progress

* constant parsing in progress

* constant values loading and dumping

* more unit testing

* Dimping instance parameters

* Collecting instance parameters

* Initiating assigmnent support

* use latest naja-verilog

* constants refactoring and testing

* More constant testing

* verilog dumper tests are back

* Enhance testing

* test constants

* create nets associated to ports in verilog

* test instance params

* Refactoring Instance Parameters

* leak cleaning

* Error management with verilog parsing

* Improve message error

* restructure tests

* Restructure testing

* Enhance error messaging

* Error management

* testing constant connectivity

* Use latest naja-verilog and test second pass error

* Enhancing errors

* start parsing assigns

* Testing assigns

* Instance parameter testing

* Testing constant values

* setTermNet testing

* restructuring verilog errors

* findTop testing

* enhance error message

* forgotten catch

* test instance parameters

* Python interface error detection

* python interface error testing

+++ src/snl/snl/kernel/SNLDesign.cpp
@@ -201,7 +201,7 @@ void SNLDesign::addTerm(SNLTerm* term) {
       flatID = scalarTerm->getFlatID() + 1;
     } else {
       SNLBusTerm* busTerm = static_cast<SNLBusTerm*>(lastTerm);
-      flatID = busTerm->flatID_ + busTerm->getSize();
+      flatID = busTerm->flatID_ + static_cast<size_t>(busTerm->getSize());
     }
     term->setFlatID(flatID);
   }
